+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 
 #### Version 8.1.0
 * Add support for PHP-FPM workload profiles
+* Add Kanboard and Blesta web app installer for Virtualmin Pro users
 * Fix PHP-FPM not restarting when certain custom config options are used in the template
+* Fix domain expiry notification being stuck in certain cases
 * Fix the missing message when DKIM is enabled or disabled
 
 #### Version 8.0.1